Add toolchain files and beman-submodule
This pull request adds a new script, beman-submodule, which provides
some of the features of git submodule but without git submodule's
disadvantage that users need to be aware of it to clone the
repository. More information can be found in
tools/beman-submodule/README.md.

It also pulls in the toolchain files from exemplar, preserving
history, using a git filter-branch.